rework Silver string building and add node:net client shims
- replace deep rope-heavy string appends with a builder-backed concat path
- teach the VM, compiler, JIT glue, and GC to read, flush, and preserve builder-backed strings
- keep arguments/object slot syncing and related frame semantics aligned with the new slot behavior
- finish fetch-backed node:http/node:https client shims and expose node:url.parse
- add regression tests for string append loops, HTTP(S) shims, and node:url exports

+++ include/pool.h
@@ -58,7 +58,7 @@ typedef struct ant_large_string_alloc {
 
 typedef struct {
   ant_offset_t len;
-  uint8_t depth;
+  uint16_t depth;
   ant_value_t left;
   ant_value_t right;
   ant_value_t cached;
